requirements.txt
streamlit==1.38.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1
pandas==2.2.2
pydantic==2.9.2
langchain==0.3.0
langchain_chroma==0.1.4
langchain_together==0.2.0
langchain_community==0.3.0
langchain_mistralai==0.2.0